feat(dashboard): build the market read banner the canvas specifies (#587)
What changed: the terminal dashboard's first main-column slot rendered
<MarketRead />, the production conditions gauge - a 58/100 score bar, five
metric boxes (Order Wall / Fear & Greed / Day / Funding / Smart Money) and
a funding override control. None of that is in Dashboard 2a.dc.html or in
specs/dashboard-2a.md. The canvas draws one eyebrow line, one 24px verdict
string and one sentence. Replaced with TMarketReadBanner, built to the
canvas geometry read straight off the frame: --bg2 ground, padding 15px
20px, bottom hairline; eyebrow 10px mono .16em uppercase --txt3; verdict
24px mono 700 line-height 1; sub 12.5px/1.5 --txt2 capped at 680px.

It calls the same computeMarketRead() the gauge called, so the verdict and
sentence are the production read rather than a second opinion - only the
presentation changes. MarketRead.tsx's terminal eyebrow branch (added when
this slot still rendered it) is now unreachable and has been removed along
with its formatUtcStamp helper, rather than left as dead design-specific
code in a component the current design still shares.

Why: the owner compared the live dashboard against the canvas band by band
and found the mismatch; QA confirmed it against the frame and the spec and
retracted dashboard's "done" verdict. Everything after this panel already
matched canvas order - this was a one-panel gap, but it is the first thing
on the route. It survived because this branch restyled the component that
was already in the slot instead of checking what the canvas wanted to
occupy it, and because spec-conformance's C1 counts sections in position
rather than which component fills each position.

Test: open /dashboard?design=terminal. The top of the main column should be
three lines - "Market read · <DD Mon HH:MM UTC>", a large verdict string,
and one sentence under it - with no score bar, no metric boxes and no
funding override. Everything below it (Best Setup Today, selected-coin
card, 6 signal cards, Next Events / Market Conditions) is unchanged. Then
open /dashboard WITHOUT the flag: the current design must still show the
full gauge with its Tip eyebrow, unchanged by this PR.

Risk: Medium. Touches a component the current design also renders
(MarketRead.tsx), though only to delete a branch terminal no longer
reaches - current-design output is byte-identical. One deliberate departure
from criterion 5, flagged for design rather than resolved unilaterally:
the spec asks the verdict to take the read's DIRECTION (bullish --green,
bearish --red), and the canvas fixture shows a directional string
("RISK-ON, CAUTIOUS"), but computeMarketRead does not produce a direction -
its band is trade-conditions QUALITY (good/mid/weak), and no market-wide
directional read exists anywhere in the codebase (confluence and Grok are
both per-coin). Colouring "Weak setup - better to wait" --red would tell
the reader the market is falling when it means the setup is poor, which is
the substitution the spec's own "Colour is data" section forbids. Mapped on
its own axis instead - favourable --green, everything else --txt2 - and
raised on #587 for design to confirm or to specify a directional read as
its own data question.
